Déplacement des données "informationsComplementaires" dans le formulaire "déclaration"
