+++ packages/app/src/components/Menu.tsx
@@ -72,7 +72,6 @@ interface Props {
   indicateurCinqFormValidated: FormState;
   informationsEntrepriseFormValidated: FormState;
   informationsDeclarantFormValidated: FormState;
-  informationsComplementairesFormValidated: FormState;
   declarationFormValidated: FormState;
 }
 
@@ -88,7 +87,6 @@ function Menu({
   indicateurCinqFormValidated,
   informationsEntrepriseFormValidated,
   informationsDeclarantFormValidated,
-  informationsComplementairesFormValidated,
   declarationFormValidated
 }: Props) {
   const width = useColumnsWidth(2);
@@ -185,11 +183,6 @@ function Menu({
                 to={`/simulateur/${code}/informations-declarant`}
                 title="informations déclarant"
                 valid={informationsDeclarantFormValidated}
-              />
-              <CustomNavLink
-                to={`/simulateur/${code}/informations-complementaires`}
-                title="informations complémentaires"
-                valid={informationsComplementairesFormValidated}
               />
               <CustomNavLink
                 to={`/simulateur/${code}/declaration`}
