+++ packages/app/src/globals.d.ts
@@ -69,15 +69,12 @@ export type AppState = {
     tel: string;
     email: string;
   };
-  informationsComplementaires: {
-    formValidated: FormState;
-    dateConsultationCSE: string;
-    datePublication: string;
-    lienPublication: string;
-  };
   declaration: {
     formValidated: FormState;
     mesuresCorrection: string;
+    dateConsultationCSE: string;
+    datePublication: string;
+    lienPublication: string;
     dateDeclaration: string;
   };
 };
@@ -210,14 +207,6 @@ export type ActionType =
       valid: FormState;
     }
   | {
-      type: "updateInformationsComplementaires";
-      data: ActionInformationsComplementairesData;
-    }
-  | {
-      type: "validateInformationsComplementaires";
-      valid: FormState;
-    }
-  | {
       type: "updateDeclaration";
       data: ActionDeclarationData;
     }
@@ -312,14 +301,11 @@ export type ActionInformationsDeclarantData = {
   email: string;
 };
 
-export type ActionInformationsComplementairesData = {
+export type ActionDeclarationData = {
+  mesuresCorrection: string;
   dateConsultationCSE: string;
   datePublication: string;
   lienPublication: string;
-};
-
-export type ActionDeclarationData = {
-  mesuresCorrection: string;
 };
 
 ////
