--- packages/app/src/views/InformationsComplementaires/index.js
@@ -1,2 +1,0 @@
-import InformationsComplementaires from "./InformationsComplementaires";
-export default InformationsComplementaires;
